scheduler, conversation: review follow-ups on the stuffed-group seal
The export decision is recurrent_export_asked(ephemeral, records_recurrent_state),
beside resolve_seal_range and used at the seal site, with a raw-case test of all
four rows: only a seal that owns the state on a timeline that keeps it asks.

When a stuffed group's last region does not seal but earlier ones did, a warning
now says so: no memory record names the group, and a resume installs the newest
record from before it. The export-failure warning said a resume would recompute
from zeros; snapshot_within_recovered_history rejects only a record that is too
new, so a resume installs the newest earlier one, behind the turn's K/V. The
warning and the eviction comment now say that.

The "block-diagonal / every case masked to itself" claims in
submit_prefilled_turn_group and npcd's dream docs were false: no mask exists, the
grid is one causal prefill, and on a hybrid the recurrence carries earlier cases
forward too. The docs now say what block alignment guarantees (no neighbour's
tokens in an exemplar's sign(Q) window) and what it does not (context). The
SealAction::TurnGroup doc said every region but the last names SealEnd::At; all
do. The restart test's comment on sealed_memory_at described a panic the helper
does not raise, and its two `as u32` casts were on a value that already is one
(clippy unnecessary_cast).

Test: a_seal_asks_for_recurrent_state_only_when_it_owns_and_keeps_it.

+++ zend/tests/restart_turn_recovery.rs
@@ -157,9 +157,10 @@ fn a_stuffed_group_records_its_memory_at_the_last_case() {
     let sealed_turns = engine.conversation().read().turn_count(timeline);
     assert_eq!(sealed_turns, 4, "one sealed turn per case");
     // Polls until the record for exactly this turn is readable, and panics if
-    // the newest record names any other turn.
-    let record = sealed_memory_at(&engine, timeline, (sealed_turns - 1) as u32);
-    assert_eq!(record.turn_index, (sealed_turns - 1) as u32);
+    // none has appeared within its window — before the fix the only record named
+    // turn 0, so this waited out the window and failed.
+    let record = sealed_memory_at(&engine, timeline, sealed_turns - 1);
+    assert_eq!(record.turn_index, sealed_turns - 1);
 }
 
 /// **A resumed conversation's new turns survive the next restart too.**
